state: removed TestingInitialize & moved NewMongoInfo and NewDialOpts to testing
